# src/main/resources/swagger/v1.json
{
  "openapi": "3.1.0",
...
  "info": {
    "title": "Mais informações sobre a API",
    "version": "1.0.0",
    "description": "API de Eventos feita com Quarkus \n - Autor: [Lucas S. Campos](https://www.linkedin.com/in/lucassantosc/) \n - Data: 06/04/2025 \n - Versão: 1.0.0 \n - Repositório: [API Eventos TSI](https://github.com/GiltMonster/API-Eventos-java) \n - ⚠️ **Atenção**: Esta versão da API ira se tornar depreciada em breve, por favor, utilize a versão mais recente da API. \n - ⚠️ **Atenção**: A V1 dessa API não possui autenticação, portanto, qualquer pessoa pode acessar os endpoints.",
    "termsOfService": "https://exemplo.com/termos",
    "contact": {
      "name": "Dev API - Support",
      "email": "lucas.scampos4@senacsp.edu.br"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "components": {
    "schemas": {
      "AtualizarSenhaUsuario": {
        "type": "object",
        "properties": {
          "senhaAtual": { "type": "string" },
          "novaSenha": { "type": "string" }
        }
      },
      "AtualizarUsuarioDTO": {
        "type": "object",
        "properties": {
          "nome": { "type": "string" },
          "sobreNome": { "type": "string" },
          "email": { "type": "string" }
        }
      },
...
      "LoginRequest": {
        "type": "object",
        "properties": {
          "email": { "type": "string" },
          "senha": { "type": "string" }
        }
      },
...
    }
  },
  "paths": {
...
    "/v1/usuarios/atualizarSenha/{id}": {
      "put": {
        "summary": "Atualizar senha do usuário",
        "description": "Essa rota é responsável por atualizar a senha de um usuário no sistema.Se o usuário for encontrado, retorna 200.Se o usuário não for encontrado, retorna 404.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "format": "int32" }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/AtualizarSenhaUsuario" }
            }
          },
          "required": true
        },
        "responses": {
          "200": { "description": "OK" },
          "400": { "description": "Bad Request" }
        },
        "tags": ["Usuario Resource"]
      }
    },
    "/v1/usuarios/atualizarUsuario/{id}": {
      "put": {
        "summary": "Atualizar usuário",
        "description": "Essa rota é responsável por atualizar um usuário no sistema.Se o usuário for encontrado, retorna 200.Se o usuário não for encontrado, retorna 404.",
        "parameters": [
          {
            "description": "ID do usuário a ser atualizado",
            "example": 1,
            "required": true,
            "name": "id",
            "in": "path",
            "schema": { "type": "integer", "format": "int32" }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/AtualizarUsuarioDTO" }
            }
          },
          "required": true
        },
        "responses": {
          "200": { "description": "OK" },
          "400": { "description": "Bad Request" }
        },
        "tags": ["Usuario Resource"]
      }
    },
    "/v1/usuarios/deleteUsuario/{id}": {
      "delete": {
        "summary": "Deletar usuário",
        "description": "Essa rota é responsável por deletar um usuário do sistema.Se o usuário for encontrado, retorna 200.Se o usuário não for encontrado, retorna 404.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "format": "int32" }
          }
        ],
        "responses": {
          "200": { "description": "OK" }
        },
        "tags": ["Usuario Resource"]
      }
    },
    "/v1/usuarios/findUsuario/{id}": {
      "get": {
        "summary": "Buscar usuário por ID",
        "description": "Essa rota é responsável por buscar um usuário pelo ID.Se o ID não for encontrado, retorna 404.OBS: O ID é gerado automaticamente pelo banco de dados, é unico e é numérico.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "format": "int32" }
          }
        ],
        "responses": {
          "200": { "description": "OK" }
        },
        "tags": ["Usuario Resource"]
      }
    },
    "/v1/usuarios/login": {
      "post": {
        "summary": "Realiza o login do usuário",
        "description": "Verifica se o usuário está cadastrado e retorna os dados do usuário",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/LoginRequest" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Login realizado com sucesso",
            "content": {
              "application/json": {
                "schema": { "type": "string" }
              }
            }
          },
          "404": { "description": "Usuário não encontrado" },
          "401": { "description": "Senha incorreta" },
          "400": { "description": "Parâmetros obrigatórios: email e senha" }
        },
        "tags": ["Usuario Resource"]
      }
    }
  }
}